requirements.txt
sentence-transformers==5.4.0
transformers==5.5.4
torch==2.11.0
faiss-cpu==1.13.2
streamlit==1.56.0
pypdf==6.10.0
python-docx==1.2.0
google-genai==1.72.0
numpy==2.4.4
python-dotenv==1.2.2
